Refactored the code for better readability and less complexity
