+++ ServiceBusDashboard/Model/SbQueue.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace ServiceBusDashboard.Model
+{
+    public class SbQueue : CountContainer
+    {
+    }
+}
