+++ ServiceBusDashboard/Views/Home/DashboardView.cshtml
@@ -1,4 +1,4 @@
-﻿@model ServiceBusDashboard.Controllers.DashboardModel
+﻿@model ServiceBusDashboard.Model.SbDashboardModel
 
 @{
     Layout = null;
