+++ ServiceBusDashboard/Web.config
@@ -6,6 +6,10 @@
 		<add key="ClientValidationEnabled" value="true" />
 		<add key="UnobtrusiveJavaScriptEnabled" value="true" />
 		<add key="SbConnectionStrings" value="..\ConnectionStrings.json" />
+		<add key="IgnoreQueues" value="" />
+		<add key="IgnoreTopics" value="" />
+		<add key="IgnoreSubscriptions" value="" />
+		<add key="ExcludeSubscriptions" value="" />
 	</appSettings>
 	<system.web>
 		<compilation debug="true" targetFramework="4.7.1" />
